fix(security): sposta getEffectivePlan fuori dal modulo "use server" (REVIEW #66)
Ogni export async di un file "use server" diventa un endpoint POST pubblico
invocabile con l'action ID estraibile dal bundle. getEffectivePlan accettava
uno userId arbitrario senza chiamare getAuthenticatedUser: info-leak del piano
altrui dato l'auth UUID, query DB gratuite non autenticate (surface DoS) e
22P02 non gestito su userId non-UUID.

È un helper server-side, non una action: spostato in src/lib/plans.ts accanto
a getPlan. Aggiornati gli import in api-key-actions.ts e nei test; i test
funzionali vivono ora in plans.test.ts, con un guard che verifica che
billing-actions.ts non esporti più il symbol.
